requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.0,<26.0
django-environ>=0.11,<1.0
psycopg[binary]>=3.1,<4.0
redis>=5.0,<6.0
celery[redis]>=5.4,<6.0
gunicorn>=22.0,<24.0
Pillow>=10.0,<12.0
django-htmx>=1.19,<2.0
django-ratelimit>=4.1,<5.0
easy-thumbnails>=2.10,<3.0
django-storages>=1.14,<2.0
whitenoise>=6.8,<7.0

